+++ Frost/src/Frost/Renderer/Buffers/IndexBuffer.h
@@ -9,8 +9,8 @@ namespace Frost
 	public:
 		virtual ~IndexBuffer() {}
 
-		virtual uint32_t GetCount() const = 0;
-		virtual uint32_t GetBufferSize() const = 0;
+		virtual uint64_t GetCount() const = 0;
+		virtual uint64_t GetBufferSize() const = 0;
 		virtual Ref<BufferDevice> GetBuffer() const = 0;
 
 		virtual void Bind() const = 0;
@@ -18,7 +18,7 @@ namespace Frost
 
 		virtual void Destroy() = 0;
 
-		static Ref<IndexBuffer> Create(void* indicies, uint32_t count);
+		static Ref<IndexBuffer> Create(void* indicies, uint64_t count);
 	};
 
 }
